[UI] Cleanup of markers + removed obsolete stuff
git-svn-id: svn://svn.berlios.de/avidemux/branches/avidemux_2.6_branch_mean@4642 30592fff-cfd4-0310-bf0e-8235a3c7be26

+++ avidemux/ADM_userInterfaces/ADM_NONE/ADM_gui2/gui_none.cpp
@@ -87,7 +87,7 @@ void 	UI_setScale( double  val )
 {}
 void 	UI_setFrameType( uint32_t frametype,uint32_t qp)
 {}
-void 	UI_setMarkers(uint32_t a, uint32_t b )
+void 	UI_setMarkers(uint64_t a, uint64_t b )
 {}
 void 	UI_setTitle(const char *name)
 {}
